requirements.txt
anyio==4.12.1
certifi==2026.2.25
dnspython==2.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
pymongo==4.16.0
python-dotenv==1.2.2
python-telegram-bot==20.7
sniffio==1.3.1
